Fix more errors
- Catch Exception class to catch all exceptions
- On Windows dont try to open unavailable pipes, this leads to a vim error

+++ python/rpc.py
@@ -145,6 +145,8 @@ class WinDiscordIpcClient(DiscordIpcClient):
     def _connect(self):
         for i in range(10):
             path = self._pipe_pattern.format(i)
+            if not os.path.exists(path):
+                continue
             try:
                 self._f = open(path, "w+b")
             except OSError as e:
